src/config/settings.py
import os
from dotenv import load_dotenv

load_dotenv()

DB_URL = os.getenv("DATABASE_URL")
KAFKA_BROKER = os.getenv("KAFKA_BROKER")
KAFKA_TOPIC = os.getenv("KAFKA_TOPIC")
STRIPE_SECRET_KEY = os.getenv("STRIPE_SECRET_KEY")
